Implement initial performance test harness
The StressTestComponent is intended to provide an ImGUI interface for
interface with any stress testing facilities/functions provided in the
future. It is provided as a network component to support RPCs and any
properties we wish to replicate that are not specific to other game
entities. Currently, it provides a single RPC to spawn N animated
entities.

There is an unknown bug preventing this component from being
instantiated on the client, so at this time, AI entities may only be
spawned on the server.

The NetworkAiComponent is attached to the player entity and disabled by
default. When enabled, the WasdMovementComponent and
NetworkWeaponsComponent on the player entity are no longer driven by
inputs. Instead, they tick and poll for the input state from the
NetworkAiComponent.

Currently, the AI component is a chaos monkey, randomizing the current
movement direction, action, and weapon state at random intervals. A
future change could set the seed and decouple the action from the
deltaTime to create reproducible tests.

- Fixed the material on the animating character
- Commits the project physmaterial
- Refactor SampleBase box entities into 4x4x4 prefabs

Signed-off-by: Jeremy Ong <jcong@amazon.com>