Update README.md
Reverting the README back to its original state. In Step 2, the steps are optional. If you register (option 1), you can then proceed to step 3. OR you can choose to use the Project Manager option, which will register the engine and gems and create the build, allowing you to skip step 3.

Signed-off-by: mcphedar <mcphedar@amazon.com>